fix: sse 를 트랜잭션에서 제외
